Quantized modules return the activation dtype: end the silent fp32 promotion cascade
MLX types quantized matmul outputs as promote_types(activation, scales).
JANG bundles stamp f16 affine scales, so on bf16 models every stock
quantized op silently returned FLOAT32 — and nothing cast back. A deep
audit of all runtime/inference code traced the cascade: fp32 residual
stream, fp32 K/V projections silently DOUBLING every KV cache and
disk-cache entry (DeepseekV3's own comment records the measured form:
46/48 layers at 2x size), and five compiled-decode dtype gates refusing
silently, so f16-scale bundles lost their fused regions with no log
(the compiled_gdn_tail=rejected ... float32 diagnostic was one visible
symptom).

Fixes, from the boundary outward:
- QuantizedLinear / QuantizedSwitchLinear pin their outputs back to the
  activation dtype when it is bf16/f16 (fp32 stays where it belongs —
  inside the kernel as the accumulator). VMLX_QUANTIZED_OUTPUT_PROMOTE=1
  restores the historical propagation for A/B.
- 21 MoE reducers get DeepseekV3's established scores.asType(y.dtype)
  pattern (Hy3 x2, Qwen3Next, Qwen35 both twins, Qwen3MoE, GLM4MOE(+Lite),
  PhiMoE, NemotronH(+JANGTQ), MiniMax(+JANGTQ), MiMoV2Flash, OlmoE,
  Mixtral, Jamba, LFM2-class via SwitchGLU, DeepseekV3/V4 JANGTQ,
  DeepseekOCR, Glm4vMoe, Qwen35JANGTQ).
- The grouped GDN input-projection fallback (both Qwen35 twins) pins its
  split parts — the unfused QuantizedLinear reference now pins the same
  way, so bit-identity is preserved (suite still 2/2).
- TurboQuantSwitchLinear casts its declared-fp32 kernel output back
  (mirroring its GLU sibling); AffineStreamingSwitchGLU (which REQUIRES
  f16 scales) casts its MoE block output; BailingHybrid's GLA layer
  returns bf16 while keeping its fp32 recurrence state; the compiled
  GDN front's conv-tail CACHE write is pinned to the activation dtype.
- NoContextScalingFP32CastSourceTests guards the explicit-cast class
  repo-wide with a justified allowlist (Phi fp16 range, Gemma-4 score
  storage, the three env-gated fallback arms).

Known remaining (documented, needs the loader module-swap lever, not a
cast): Gemma-4 jang_affine mmap and DSV4 prestacked mmap preserve f16
scales with no native-module swap, so their QuantizedEmbedding output
enters the stream as f16 and the first mixed op still promotes.

Proof: focused suite 772 tests — failing set IDENTICAL to clean main
(21 pre-existing environment flakes, zero new); fusion bit-identity 2/2,
MLA dtype 3/3, DSV4 indexer causal top-k green, source guard green.

+++ Libraries/MLXLLM/Models/BailingHybrid.swift
@@ -680,7 +680,12 @@ class BailingLinearAttention: Module {
             cache?.offset += L
         }
 
-        let flat = output.transposed(0, 2, 1, 3).reshaped(B, L, -1)
+        // The GLA recurrence intentionally runs and stores its STATE in fp32
+        // (fp16-overflow fix above); the layer OUTPUT has no such need, and
+        // returning it fp32 promoted the residual stream — and through it
+        // every downstream attention layer's KV cache — to fp32 on bf16
+        // bundles. Cast the output lane back; the state lane stays fp32.
+        let flat = output.transposed(0, 2, 1, 3).reshaped(B, L, -1).asType(x.dtype)
         let gated = gNorm(flat) * sigmoid(gProj(x))
         return dense(gated)
     }
